+++ features/collections/ui/src/commonMain/kotlin/app/campfire/collections/ui/list/CollectionsPresenter.kt
@@ -5,6 +5,7 @@ import androidx.compose.runtime.collectAsState
 import androidx.compose.runtime.getValue
 import androidx.compose.runtime.remember
 import app.campfire.analytics.Analytics
+import app.campfire.analytics.events.ActionEvent
 import app.campfire.analytics.events.ContentSelected
 import app.campfire.analytics.events.ContentType
 import app.campfire.collections.api.CollectionsRepository
@@ -13,6 +14,7 @@ import app.campfire.common.screens.CollectionsScreen
 import app.campfire.core.coroutines.LoadState
 import app.campfire.core.di.UserScope
 import app.campfire.core.model.Collection
+import app.campfire.settings.api.CampfireSettings
 import com.r0adkll.kimchi.circuit.annotations.CircuitInject
 import com.slack.circuit.foundation.NonPausablePresenter
 import com.slack.circuit.runtime.Navigator
@@ -26,6 +28,7 @@ import me.tatarka.inject.annotations.Inject
 class CollectionsPresenter(
   @Assisted private val navigator: Navigator,
   private val repository: CollectionsRepository,
+  private val settings: CampfireSettings,
   private val analytics: Analytics,
 ) : NonPausablePresenter<CollectionsUiState> {
 
@@ -38,11 +41,21 @@ class CollectionsPresenter(
         .catch<LoadState<out List<Collection>>> { emit(LoadState.Error) }
     }.collectAsState(LoadState.Loading)
 
+    val displayState by remember {
+      settings.observeCollectionsDisplayState()
+    }.collectAsState()
+
     return CollectionsUiState(
       collectionContentState = collectionContentState,
+      displayState = displayState,
     ) { event ->
       when (event) {
         CollectionsUiEvent.Back -> navigator.pop()
+
+        CollectionsUiEvent.ToggleDisplayState -> {
+          analytics.send(ActionEvent("collections_display_state", "toggle"))
+          settings.collectionsDisplayState = displayState.next()
+        }
 
         is CollectionsUiEvent.CollectionClick -> {
           analytics.send(ContentSelected(ContentType.Collection))