+++ features/playlists/ui/src/commonMain/kotlin/app/campfire/playlists/ui/list/PlaylistsPresenter.kt
@@ -5,6 +5,7 @@ import androidx.compose.runtime.collectAsState
 import androidx.compose.runtime.getValue
 import androidx.compose.runtime.remember
 import app.campfire.analytics.Analytics
+import app.campfire.analytics.events.ActionEvent
 import app.campfire.analytics.events.ContentSelected
 import app.campfire.analytics.events.ContentType
 import app.campfire.core.coroutines.LoadState
@@ -13,6 +14,7 @@ import app.campfire.core.model.Playlist
 import app.campfire.playlists.api.PlaylistsRepository
 import app.campfire.playlists.api.screen.PlaylistDetailScreen
 import app.campfire.playlists.api.screen.PlaylistsScreen
+import app.campfire.settings.api.CampfireSettings
 import com.r0adkll.kimchi.circuit.annotations.CircuitInject
 import com.slack.circuit.foundation.NonPausablePresenter
 import com.slack.circuit.runtime.Navigator
@@ -26,6 +28,7 @@ import me.tatarka.inject.annotations.Inject
 class PlaylistsPresenter(
   @Assisted private val navigator: Navigator,
   private val playlistsRepository: PlaylistsRepository,
+  private val settings: CampfireSettings,
   private val analytics: Analytics,
 ) : NonPausablePresenter<PlaylistsUiState> {
 
@@ -37,11 +40,21 @@ class PlaylistsPresenter(
         .catch<LoadState<out List<Playlist>>> { emit(LoadState.Error) }
     }.collectAsState(LoadState.Loading)
 
+    val displayState by remember {
+      settings.observePlaylistsDisplayState()
+    }.collectAsState()
+
     return PlaylistsUiState(
       playlistContentState = playlistContentState,
+      displayState = displayState,
     ) { event ->
       when (event) {
         PlaylistsUiEvent.Back -> navigator.pop()
+
+        PlaylistsUiEvent.ToggleDisplayState -> {
+          analytics.send(ActionEvent("playlists_display_state", "toggle"))
+          settings.playlistsDisplayState = displayState.next()
+        }
 
         is PlaylistsUiEvent.PlaylistClick -> {
           analytics.send(ContentSelected(ContentType.Collection))